lsp_core
Module prelude

Re-exports

pub use completion::Label as CompletionLabel;

pub use hover::Label as HoverLabel;

pub use parse::Label as ParseLabel;

pub use rename::PrepareRename as PrepareRenameLabel;

pub use rename::Rename as RenameLabel;

pub use diagnostics::Label as DiagnosticsLabel;

pub use save::Label as SaveLabel;

pub use inlay::Label as InlayLabel;

pub use format::Label as FormatLabel;

pub use semantic::Label as SemanticLabel;

Modules

completion

diagnostics

feature

Hosts all common features of the semantic language server. All features supported by the language server. Features like completion (CompletionRequest) and hover expose a component that collects items returned to the client.

format

hover

inlay

parse

rename

save

semantic

semantic_token

systems

Structs

Backend

CommandReceiver

[Resource] used to receive command queues. These command queues are handled with handle_tasks.

CommandSender

[Resource] used to send command queues, allowing for async operations.

CompletionRequest

[Component] indicating that the current document is currently handling a Completion request.

CreateEvent

Event triggers when a document is opened

DiagnosticItem

DiagnosticPublisher

DiagnosticSender

Dirty

[Component] indicating whether or not the document is dirty, a dirty document parsed with errors.

DocumentLinks

[Component] used to remember the linked documents.

DynLang

[Component] that allows for language specific implementation for certain things, reducing code duplication.

Element

[Component] that contains the parsed semantic element (i.e. Turtle, JSONLD).

Errors

Simple wrapper for errors that derives [Component]

FormatRequest

[Component] indicating that the current document is currently handling a Format request.

HighlightRequest

[Component] indicating that the current document is currently handling a Hightlight request.

HoverRequest

[Component] indicating that the current document is currently handling a Hover request.

InlayRequest

[Component] indicating that the current document is currently handling a Inlay request.

KeyWords

[Component] containing the typical keywords for the current language.

Label

[Component] containing the lsp_types::Url of the current document.

MyQuad

MyTerm

Open

[Component] indicating whether or not the document is actually open.

PositionComponent

[Component] used to wrap an incoming [lsp_types::Position].

Prefix

One defined prefix, maps prefix to url

Prefixes

[Component] that containing defined prefixes and base URL.

PrepareRenameRequest

[Component] indicating that the current document is currently handling a PrepareRename request.

RenameEdits

[Component] indicating that the current document is currently handling a Rename request, collecting TextEdits.

RopeC

[Component] containing the current source code as [ropey::Rope]

SemanticTokensDict

[Resource] mapping a [‘SemanticTokenType’] to their used index.

SimpleCompletion

SimpleDiagnostic

Source

[Component] containing the current source code as String

Spanned

TokenComponent

[Component] used to indicate the currently targeted Token during a request.

Tokens

[Component] that contains the parsed tokens.

TripleComponent

[Component] used to indicate the currently targeted [MyQuad<'static>] during a request.

Triples

[Component] containing all derived Triples from the documents.

Triples2

TypeHierarchy

[Resource] used to set and get all super and subtypes starting from a TypeId

Types

maps terms to all known correct types.

Wrapped

Simple wrapper structure that derives [Component]

Enums

SparqlAggregate

SparqlCall

SparqlExpr

SparqlExpr2

SparqlExpr3

SparqlKeyword

StringStyle

Token

TripleTarget

[Component] used to indicate the term type of currently targeted Token in the Triple.

Traits

Client

ClientSync

Lang

LangHelper

Membered

TokenTrait

Functions

get_current_token

get_current_triple

lsp_range_to_range

offset_to_position

offsets_to_range

position_to_offset

range_to_range

setup_schedule_labels

Initializes a [World], including Resources and [Schedules]. All systems defined in crate are added to the [World].

spanned

spawn_or_insert