lsp_core::prelude
Module feature

Hosts all common features of the semantic language server. All features supported by the language server. Features like completion (CompletionRequest) and hover expose a component that collects items returned to the client.

Re-exports

pub use completion::Label as CompletionLabel;

pub use hover::Label as HoverLabel;

pub use parse::Label as ParseLabel;

pub use rename::PrepareRename as PrepareRenameLabel;

pub use rename::Rename as RenameLabel;

pub use diagnostics::Label as DiagnosticsLabel;

pub use save::Label as SaveLabel;

pub use inlay::Label as InlayLabel;

pub use format::Label as FormatLabel;

pub use semantic::Label as SemanticLabel;

Modules

completion

diagnostics

format

hover

inlay

parse

rename

save

semantic