lsp_core::prelude
Module parse

Re-exports

pub use crate::systems::derive_classes;

pub use crate::systems::derive_prefix_links;

pub use crate::systems::derive_properties;

pub use crate::systems::derive_shapes;

pub use crate::systems::extract_type_hierarchy;

pub use crate::systems::fetch_lov_properties;

pub use crate::systems::infer_types;

Structs

Label

ScheduleLabel related to the Parse schedule

Functions

prefixes

Parse schedule barrier, after this system, prefixes should be derived

setup_schedule

triples

Parse schedule barrier, after this system, triples should be derived