lsp_core::prelude

Module rename

Re-exports

pub use crate::util::token::get_current_token;

Structs

PrepareRename

ScheduleLabel related to the PrepareRename schedule

PrepareRenameRequest

Component indicating that the current document is currently handling a PrepareRename request.

Rename

ScheduleLabel related to the Rename schedule

RenameEdits

Component indicating that the current document is currently handling a Rename request, collecting TextEdits.

Functions

prepare_rename

rename

setup_schedules