lsp_core::prelude

Module systems

Modules

prefix

Structs

DefinedClass

DefinedProperty

FromPrefix

OntologyExtractor

TypeId

Functions

check_added_ontology_extract

complete_class

complete_properties

derive_classes

derive_owl_imports_links

derive_prefix_links

derive_properties

derive_shapes

extract_type_hierarchy

fetch_lov_properties

First of al, fetch the lov dataset information at url https://lov.linkeddata.es/dataset/lov/api/v2/vocabulary/info?vocab=${prefix} Next, extract that json object into an object and find the latest dataset

handle_tasks

hover_class

hover_property

hover_types

infer_current_type

infer_types

init_onology_extractor

inlay_triples

keyword_complete

open_imports

spawn_or_insert

validate_shapes

System evaluates linked shapes

validate_with_updated_shapes

System checks what entities should retrigger a shape evaluation when a shape changes

Type Aliases

DefinedClasses

DefinedProperties