lsp_core::prelude::systems

Function fetch_lov_properties

pub fn fetch_lov_properties<C: Client + Resource>(
    sender: Res<'_, CommandSender>,
    query: Query<'_, '_, &Prefixes, (Or<((Changed<Prefixes>, With<Open>), Changed<Open>)>,)>,
    ontologies: Query<'_, '_, (Entity, &LocalPrefix)>,
    prefixes: Local<'_, HashSet<String>>,
    client: Res<'_, C>,
    fs: Res<'_, Fs>,
)

First of al, fetch the lov dataset information at url https://lov.linkeddata.es/dataset/lov/api/v2/vocabulary/info?vocab=${prefix} Next, extract that json object into an object and find the latest dataset