lsp_core
Module systems

Modules

prefix

Structs

DefinedClass

DefinedProperty

TypeId

Functions

complete_class

complete_properties

derive_classes

derive_owl_imports_links

derive_prefix_links

derive_properties

derive_shapes

extract_type_hierarchy

fetch_lov_properties

First of al, fetch the lov dataset information at url https://lov.linkeddata.es/dataset/lov/api/v2/vocabulary/info?vocab=${prefix} Next, extract that json object into an object and find the latest dataset

handle_tasks

hover_class

hover_property

hover_types

infer_types

inlay_triples

keyword_complete

spawn_or_insert

validate_shapes

System evaluates linked shapes

validate_with_updated_shapes

System checks what entities should retrigger a shape evaluation when a shape changes