lsp_core
Module components

Defines all common [Component]s and [Resource]s

In this [World], [Entity]s are documents and Components are derived from these documents. Different [System]s derive new Components from existing Components, that are added to the [Entity]. For example, if Triples are defined, systems::derive_classes will derive DefinedClass from them and add them to the [Entity].

Structs

CommandReceiver

[Resource] used to receive command queues. These command queues are handled with handle_tasks.

CommandSender

[Resource] used to send command queues, allowing for async operations.

Dirty

[Component] indicating whether or not the document is dirty, a dirty document parsed with errors.

DocumentLinks

[Component] used to remember the linked documents.

DynLang

[Component] that allows for language specific implementation for certain things, reducing code duplication.

Element

[Component] that contains the parsed semantic element (i.e. Turtle, JSONLD).

Errors

Simple wrapper for errors that derives [Component]

KeyWords

[Component] containing the typical keywords for the current language.

Label

[Component] containing the lsp_types::Url of the current document.

Open

[Component] indicating whether or not the document is actually open.

PositionComponent

[Component] used to wrap an incoming [lsp_types::Position].

RopeC

[Component] containing the current source code as [ropey::Rope]

Source

[Component] containing the current source code as String

TypeHierarchy

[Resource] used to set and get all super and subtypes starting from a TypeId

Types

maps terms to all known correct types.

Wrapped

Simple wrapper structure that derives [Component]